England lost final in Euro 2021. Commit: upgraded routes middleware, autoload in postfeeds and fixed request validation

+++ routes/web.php
@@ -3,7 +3,6 @@
 use App\Http\Controllers\AccountUserPermissionController;
 use App\Http\Controllers\FriendController;
 use App\Http\Controllers\PostController;
-use App\Http\Controllers\UploadImageController;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\UserController;
 use App\Http\Controllers\AccountController;
@@ -22,8 +21,6 @@ use App\Http\Controllers\KidController;
 //****************************************************************************************************************************
 // Images routes
 //****************************************************************************************************************************
-Route::post('/image/account',       [UploadImageController::class, 'save'])->name('account.avatar')->middleware('auth');
-
 
 
 Route::get('/',                     [PostController::class, 'feedPosts'])->name('welcome')->middleware('auth');
@@ -32,36 +29,42 @@ Route::get('/',                     [PostController::class, 'feedPosts'])->name(
 //****************************************************************************************************************************
 // Post routes
 //****************************************************************************************************************************
-Route::get('/posts',        [PostController::class, 'index'])->name('posts.index')->middleware('auth');
-Route::get('/postsfeed',    [PostController::class, 'feedPosts'])->name('postsfeed.index')->middleware('auth');
-Route::get('/post/create',  [PostController::class, 'create'])->name('posts.create')->middleware('auth');
-Route::post('/posts',       [PostController::class, 'store'])->name('posts.store')->middleware('auth');
+Route::get('/posts',            [PostController::class, 'index'])->name('posts.index')->middleware('auth','admin');
+Route::get('/postsfeed',        [PostController::class, 'feedPosts'])->name('postsfeed.index')->middleware('auth');
+Route::get('/post/create',      [PostController::class, 'create'])->name('posts.create')->middleware('auth','publisher');
+Route::post('/posts',           [PostController::class, 'store'])->name('posts.store')->middleware('auth','publisher');
+Route::get('/posts/edit/{post}',[PostController::class, 'edit'])->name('posts.edit')->middleware('auth','publisher');
+Route::delete('/posts/{post}',  [PostController::class, 'destroy'])->middleware('auth','publisher');
+Route::post('/posts/{post}',    [PostController::class, 'update'])->name('posts.update')->middleware('auth','publisher');
 
 
 //****************************************************************************************************************************
 // User routes
 //****************************************************************************************************************************
-Route::get('/users/list',       [UserController::class, 'index'])->name('users.index')->middleware('auth');
-Route::delete('/users/{user}',  [UserController::class, 'destroy'])->middleware('auth');
+Route::get('/users/list',       [UserController::class, 'index'])->name('users.index')->middleware('auth','admin');
+Route::delete('/users/{user}',  [UserController::class, 'destroy'])->name('users.destroy')->middleware('auth','admin');
 
 
 //****************************************************************************************************************************
 // Kid routes
 //****************************************************************************************************************************
-Route::get('/kids',             [KidController::class, 'index'])->name('kids.index')->middleware('auth');
-Route::get('/kid/create',       [KidController::class, 'create'])->name('kids.create')->middleware('auth');
-Route::post('/kids',            [KidController::class, 'store'])->name('kids.store')->middleware('auth');
+Route::get('/kids',             [KidController::class, 'index'])->name('kids.index')->middleware('auth','has_kids');
+Route::get('/kid/create',       [KidController::class, 'create'])->name('kids.create')->middleware('auth','parent');
+Route::post('/kids',            [KidController::class, 'store'])->name('kids.store')->middleware('auth','parent');
+Route::get('/kids/edit/{kid}',  [KidController::class, 'edit'])->name('kids.edit')->middleware('auth','parent');
+Route::post('/kids/{kid}',      [KidController::class, 'update'])->name('kids.update')->middleware('auth','parent');
+Route::delete('/kids/{kid}',    [KidController::class, 'destroy'])->name('kids.destroy')->middleware('auth','parent');
 
 //****************************************************************************************************************************
 // Account routes
 //****************************************************************************************************************************
 Route::get('/accounts',                 [AccountController::class, 'index'])->name('accounts.index')->middleware('auth');
-Route::get('/account/create',          [AccountController::class, 'create'])->name('accounts.create')->middleware('auth');
+Route::get('/account/create',           [AccountController::class, 'create'])->name('accounts.create')->middleware('auth','accountless');
 Route::get('/accounts/{account}',       [AccountController::class, 'show'])->name('accounts.show')->middleware('auth');
-Route::post('/accounts',                [AccountController::class, 'store'])->name('accounts.store')->middleware('auth');
-Route::get('/accounts/edit/{account}',  [AccountController::class, 'edit'])->name('accounts.edit')->middleware('auth');
-Route::post('/accounts/{account}',      [AccountController::class, 'update'])->name('accounts.update')->middleware('auth');
-Route::delete('/accounts/{account}',    [AccountController::class, 'destroy'])->name('accounts.destroy')->middleware('auth');
+Route::post('/accounts',                [AccountController::class, 'store'])->name('accounts.store')->middleware('auth','accountless');
+Route::get('/accounts/edit/{account}',  [AccountController::class, 'edit'])->name('accounts.edit')->middleware('auth','parent');
+Route::post('/accounts/{account}',      [AccountController::class, 'update'])->name('accounts.update')->middleware('auth','parent');
+Route::delete('/accounts/{account}',    [AccountController::class, 'destroy'])->name('accounts.destroy')->middleware('auth','parent');
 
 //****************************************************************************************************************************
 // AccountUserPermission routes
